Rename app-sdk script
